+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -78,5 +78,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-08-12T12:59:35.165Z")
+    updated: Optional[datetime] = parse_datetime("2025-08-13T06:47:46.437Z")
     r"""The date and time the resource was last updated"""